filter on cypher fields in relationship properties
