--- packages/graphql/tests/integration/directives/cypher/cypher-in-relationship-props.int.test.ts
@@ -1,107 +1,0 @@
-/*
- * Copyright (c) "Neo4j"
- * Neo4j Sweden AB [http://neo4j.com]
- *
- * This file is part of Neo4j.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import type { UniqueType } from "../../../utils/graphql-types";
-import { TestHelper } from "../../../utils/tests-helper";
-
-describe("cypher directive in relationship properties", () => {
-    const testHelper = new TestHelper();
-
-    let Movie: UniqueType;
-    let Actor: UniqueType;
-
-    beforeEach(async () => {
-        Movie = testHelper.createUniqueType("Movie");
-        Actor = testHelper.createUniqueType("Actor");
-
-        const typeDefs = /* GraphQL */ `
-            type ${Movie} @node {
-                title: String!
-                actors: [${Person}!]! @relationship(type: "ACTED_IN", direction: IN, properties: "ActedIn")
-            }
-
-            type ${Person} @node {
-                name: String!
-                actedIn: [${Movie}!]! @relationship(type: "ACTED_IN", direction: OUT, properties: "ActedIn")
-            }
-
-            type ActedIn @relationshipProperties {
-                screenTime: Int
-                    @cypher(
-                        statement: """
-                        RETURN 10 as c
-                        """
-                        columnName: "c"
-                    )
-            }
-        `;
-
-        await testHelper.initNeo4jGraphQL({ typeDefs });
-    });
-
-    afterEach(async () => {
-        await testHelper.close();
-    });
-
-    test("should query custom query and return relationship properties", async () => {
-        const source = /* GraphQL */ `
-            query {
-                ${Movie.plural} {
-                    title
-                    actorsConnection {
-                        edges {
-                            properties {
-                                screenTime
-                            }
-                            node {
-                                name
-                            }
-                        }
-                    }
-                }
-            }
-        `;
-
-        await testHelper.executeCypher(`CREATE(:Movie {title: "Da Matrix"})<-[:ACTED_IN]-(:Person {name: "Keanu"})`);
-
-        const gqlResult = await testHelper.executeGraphQL(source);
-
-        expect(gqlResult.errors).toBeFalsy();
-
-        expect(gqlResult.data).toEqual({
-            [Movie.plural]: [
-                {
-                    title: "Da Matrix",
-                    actorsConnection: {
-                        edges: [
-                            {
-                                properties: {
-                                    screenTime: 10,
-                                },
-                                node: {
-                                    name: "Keanu",
-                                },
-                            },
-                        ],
-                    },
-                },
-            ],
-        });
-    });
-});
